requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
python-dotenv==1.0.0
mysqlclient==2.2.0
WTForms==3.0.1
email-validator==2.0.0
werkzeug==2.3.7
Pillow==10.0.0
python-dateutil==2.8.2
email_validator
Pillow
chart.js==0.1.0
fullcalendar==5.11.3